requirements.txt
argh==0.26.2
comtypes==1.1.7
enum34==1.1.6
future==0.18.2
infi.systray==0.1.12
pathtools==0.1.2
psutil==5.6.7
pycaw==20181226
pyserial==3.4
PyYAML==5.3
watchdog==0.9.0
